requirements.txt
starlette>=0.27.0
uvicorn>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
html5lib>=1.1
markdown>=3.5
pydantic>=1.8.0,<3.0.0
trafilatura>=1.8.0,<2.0.0
python-dateutil>=2.8.0,<3.0.0
cachetools>=5.0.0,<6.0.0
filetype>=1.1.0,<2.0.0
pymupdf>=1.20.0,<2.0.0
pymupdf4llm>=0.0.20,<0.1
tzdata>=2022.1
scikit-learn>=1.3.0
charset-normalizer>=3.3.0
